+++ test/cases/require-styles-from-module.ext/webpack.config.js
@@ -1,0 +1,43 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+//const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+
+module.exports = {
+  output: {
+    path: path.join(__dirname, 'public/'),
+  },
+
+  entry: {
+    index: './src/views/index.pug',
+    //styles: './src/views/styles.scss',
+  },
+
+  plugins: [
+    //new MiniCssExtractPlugin(),
+    new PugPlugin(),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+        //use: [MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader'],
+      },
+      {
+        test: /\.(png|jpg|jpeg|ico)/,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+};